問問題方法最終抽象化 一大堆事件訂閱包裝在 Svc, BaseModule 去找 Svc 和 Helper 湊成訊息一來一往, Module 只需要呼叫一個方法就可以達成

+++ DiscordBot-MagirecoTW/Services/ResponsiveService.cs
@@ -94,7 +94,7 @@ namespace MitamaBot.Services
         public async Task<int?> WaitForNumberAnswerAsync(
             ulong channelId,
             List<Emoji> optionEmojis,
-            byte startNumber = 1,
+            int startNumber = 1,
             bool isCancellable = false,
             TimeSpan? expireAfter = null,
             TaskCompletionSource<SocketMessageOrReaction> tcs = null)
